+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 CHANGELOG
 
 ## vNext
+- Added _C#_ benchmarking types for `ISourceGenerator`.
 
 ## v0.1.0 (2022-07-24)
 - Added _C#_ benchmarking types for `IIncrementalGenerator`.